Workaround for hardcoded IPs in forwarder instance
In vpc/rdns.tf, a private IP address is hardcoded to the base of the
subnet's CIDR block with an offset of 4:

    # First four IP addresses on each subnet (counting the network address) are
    # reserved by AWS, so compute the fifth one.  If that IP is already taken,
    # you can specify a different one instead, but this should usually work.
    private_ip = cidrhost(module.public-facing-subnet["public1-a-net"].cidr_block, 4)

A comment in modules/rdns-forwarder/module.tf states that this is done
intentionally:

    # explicitly specifying private_ip helps minimize disruption to other services
    # when we destroy and re-create the RDNS Forwarder later on

Unfortunatly, changing a existing VPC and subnets to use rdns_option = 3
breaks our configurtion because, by sheer happenstance, one of the
magically-computed IP addresses happened to clash with the address
previously assigned to our Transit Gateway, which is effectively immovable.

We hacked this by changing the offset of 4 to 16 and rerunning. This time the
new magically-computed IP address was available, and we're off and running.

It might be worth looking at other options for assigning IP addresses that
are "sticky", like using an EIP. If we come up with such a solution, we'll
put it into a PR for the module's author to review.

In the meantime, adding "private_ip" to the ignore_changes block for the
forwarder allows a clean `terraform plan` after removing our temporary
hack of the IP address assignment code.